moved eval functions to test
